nuvlinkコンポーネントの切り出し
