+++ package.json
@@ -12,15 +12,13 @@
     "@fortawesome/free-brands-svg-icons": "^5.15.4",
     "@fortawesome/free-solid-svg-icons": "^5.15.4",
     "@fortawesome/react-fontawesome": "^0.1.16",
-    "@types/react-scroll": "^1.8.3",
     "classnames": "^2.3.1",
     "emailjs-com": "^3.2.0",
     "next": "12.0.3",
     "react": "17.0.2",
     "react-dom": "17.0.2",
     "react-intersection-observer": "^8.32.4",
-    "react-responsive-carousel": "^3.2.22",
-    "react-scroll": "^1.8.4"
+    "react-responsive-carousel": "^3.2.22"
   },
   "devDependencies": {
     "@types/node": "16.11.7",
